add SearchResultFragment for search event
Reviewed by Wizley.
2020.05.06.

I have added new Fragment to show results of query execution with EditText's input. For efficiency, have created SearchCard type to apply to Recycler view and SearchRecyclerAdapter helps it to be created. I tried constraintLayout to inflate card layout but only one result shown to device. So instead, used linearlayout with tablelayout to solve this prob.